Revert "Let's use int32_t instead of double"
This reverts commit 4f50ce7f1599242a181a3c8f6ab2c48e23eaa7b1.

+++ src/hb-ot-layout-common-private.hh
@@ -795,7 +795,7 @@ struct Device
 
     if (!pixels) return 0;
 
-    return pixels * (int32_t) scale / ppem;
+    return (int) (pixels * (double) scale / ppem);
   }
